refactor: refactor and resuse EGASP code for AMC substance upload and validation
