+++ src/domain/usecases/data-entry/amc/ImportAMCProductLevelData.ts
@@ -14,7 +14,7 @@ import { D2TrackerEnrollment, D2TrackerEnrollmentAttribute } from "@eyeseetea/d2
 import { D2TrackerEvent } from "@eyeseetea/d2-api/api/trackerEvents";
 import { mapResponseToImportSummary } from "../amr-individual-funghi/ImportRISIndividualFunghiFile";
 import { getStringFromFile } from "../utils/fileToString";
-import { readTemplate } from "../egasp/ImportEGASPFile";
+import { readTemplate } from "../ImportBLTemplateEventProgram";
 
 export const AMC_PRODUCT_REGISTER_PROGRAM_ID = "G6ChA5zMW9n";
 const AMR_RAW_PRODUCT_CONSUMPTION_STAGE_ID = "GmElQHKXLIE";
@@ -43,96 +43,97 @@ export class ImportAMCProductLevelData {
 
             return this.instanceRepository.getProgram(AMC_PRODUCT_REGISTER_PROGRAM_ID).flatMap(amcProgram => {
                 if (amcTemplate) {
-                    return readTemplate(amcTemplate, amcProgram, this.excelRepository, this.instanceRepository).flatMap(
-                        dataPackage => {
-                            if (dataPackage) {
-                                if (action === "CREATE_AND_UPDATE") {
-                                    //TO DO : Validate data  - Org unit, period , any other ?
+                    return readTemplate(
+                        amcTemplate,
+                        amcProgram,
+                        this.excelRepository,
+                        this.instanceRepository,
+                        AMC_PRODUCT_REGISTER_PROGRAM_ID
+                    ).flatMap(dataPackage => {
+                        if (dataPackage) {
+                            if (action === "CREATE_AND_UPDATE") {
+                                //TO DO : Validate data  - Org unit, period , any other ?
 
-                                    return this.mapAMCProductDataToTrackedEntities(
-                                        dataPackage,
-                                        orgUnitId,
-                                        orgUnitName
-                                    ).flatMap(entities => {
-                                        return this.trackerRepository
-                                            .import({ trackedEntities: entities }, action)
-                                            .flatMap(response => {
-                                                const { summary, entityIdsList } = mapResponseToImportSummary(response);
+                                return this.mapAMCProductDataToTrackedEntities(
+                                    dataPackage,
+                                    orgUnitId,
+                                    orgUnitName
+                                ).flatMap(entities => {
+                                    return this.trackerRepository
+                                        .import({ trackedEntities: entities }, action)
+                                        .flatMap(response => {
+                                            const { summary, entityIdsList } = mapResponseToImportSummary(response);
 
-                                                const primaryUploadId = localStorage.getItem("primaryUploadId");
-                                                if (entityIdsList.length > 0 && primaryUploadId) {
-                                                    //Enrollments were imported successfully, so create and uplaod a file with enrollments ids
-                                                    // and associate it with the upload datastore object
-                                                    const enrollmentIdListBlob = new Blob(
-                                                        [JSON.stringify(entityIdsList)],
-                                                        {
-                                                            type: "text/plain",
-                                                        }
-                                                    );
-
-                                                    const enrollmentIdsListFile = new File(
-                                                        [enrollmentIdListBlob],
-                                                        `${primaryUploadId}_enrollmentIdsFile`
-                                                    );
-
-                                                    return this.glassDocumentsRepository
-                                                        .save(enrollmentIdsListFile, moduleName)
-                                                        .flatMap(fileId => {
-                                                            return this.glassUploadsRepository
-                                                                .setEventListFileId(primaryUploadId, fileId)
-                                                                .flatMap(() => {
-                                                                    return Future.success(summary);
-                                                                });
-                                                        });
-                                                } else {
-                                                    return Future.success(summary);
-                                                }
-                                            });
-                                    });
-                                } else {
-                                    if (eventListId) {
-                                        return this.glassDocumentsRepository.download(eventListId).flatMap(file => {
-                                            return Future.fromPromise(getStringFromFile(file)).flatMap(_enrollments => {
-                                                const enrollmemtIdList: [] = JSON.parse(_enrollments);
-                                                const trackedEntities = enrollmemtIdList.map(id => {
-                                                    const trackedEntity: D2TrackerTrackedEntity = {
-                                                        orgUnit: orgUnitId,
-                                                        trackedEntity: id,
-                                                        trackedEntityType: AMR_GLASS_AMC_TET_PRODUCT_REGISTER,
-                                                    };
-
-                                                    return trackedEntity;
+                                            const primaryUploadId = localStorage.getItem("primaryUploadId");
+                                            if (entityIdsList.length > 0 && primaryUploadId) {
+                                                //Enrollments were imported successfully, so create and uplaod a file with enrollments ids
+                                                // and associate it with the upload datastore object
+                                                const enrollmentIdListBlob = new Blob([JSON.stringify(entityIdsList)], {
+                                                    type: "text/plain",
                                                 });
 
-                                                return this.trackerRepository
-                                                    .import({ trackedEntities: trackedEntities }, action)
-                                                    .flatMap(response => {
-                                                        const { summary } = mapResponseToImportSummary(response);
-                                                        return Future.success(summary);
+                                                const enrollmentIdsListFile = new File(
+                                                    [enrollmentIdListBlob],
+                                                    `${primaryUploadId}_enrollmentIdsFile`
+                                                );
+
+                                                return this.glassDocumentsRepository
+                                                    .save(enrollmentIdsListFile, moduleName)
+                                                    .flatMap(fileId => {
+                                                        return this.glassUploadsRepository
+                                                            .setEventListFileId(primaryUploadId, fileId)
+                                                            .flatMap(() => {
+                                                                return Future.success(summary);
+                                                            });
                                                     });
-                                            });
+                                            } else {
+                                                return Future.success(summary);
+                                            }
                                         });
-                                    } else {
-                                        //No enrollments were created during import, so no events to delete.
-                                        const summary: ImportSummary = {
-                                            status: "SUCCESS",
-                                            importCount: {
-                                                ignored: 0,
-                                                imported: 0,
-                                                deleted: 0,
-                                                updated: 0,
-                                            },
-                                            nonBlockingErrors: [],
-                                            blockingErrors: [],
-                                        };
-                                        return Future.success(summary);
-                                    }
-                                }
+                                });
                             } else {
-                                return Future.error("Cannot find data package");
+                                if (eventListId) {
+                                    return this.glassDocumentsRepository.download(eventListId).flatMap(file => {
+                                        return Future.fromPromise(getStringFromFile(file)).flatMap(_enrollments => {
+                                            const enrollmemtIdList: [] = JSON.parse(_enrollments);
+                                            const trackedEntities = enrollmemtIdList.map(id => {
+                                                const trackedEntity: D2TrackerTrackedEntity = {
+                                                    orgUnit: orgUnitId,
+                                                    trackedEntity: id,
+                                                    trackedEntityType: AMR_GLASS_AMC_TET_PRODUCT_REGISTER,
+                                                };
+
+                                                return trackedEntity;
+                                            });
+
+                                            return this.trackerRepository
+                                                .import({ trackedEntities: trackedEntities }, action)
+                                                .flatMap(response => {
+                                                    const { summary } = mapResponseToImportSummary(response);
+                                                    return Future.success(summary);
+                                                });
+                                        });
+                                    });
+                                } else {
+                                    //No enrollments were created during import, so no events to delete.
+                                    const summary: ImportSummary = {
+                                        status: "SUCCESS",
+                                        importCount: {
+                                            ignored: 0,
+                                            imported: 0,
+                                            deleted: 0,
+                                            updated: 0,
+                                        },
+                                        nonBlockingErrors: [],
+                                        blockingErrors: [],
+                                    };
+                                    return Future.success(summary);
+                                }
                             }
+                        } else {
+                            return Future.error("Cannot find data package");
                         }
-                    );
+                    });
                 } else {
                     return Future.error("Cannot find template");
                 }
